refactor to introduce an InMemoryStore that will be used in the app and in tests
