+++ 01_inventory/in_memory_store.rb
@@ -1,0 +1,17 @@
+class InMemoryStore
+  def initialize(records)
+    @records = records
+  end
+
+  def create(record)
+    records << record
+  end
+
+  def all_articles
+    records
+  end
+
+  private
+
+  attr_reader :records
+end
